Highlighting works

+++ src/components/health.rs
@@ -2,6 +2,6 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct Health {
-    pub current: u32,
-    pub max: u32,
+    pub current: f32,
+    pub max: f32,
 }
